prepare gameboard for state
