+++ src/app/_components/Gameboard/OpponentCardTray/OpponentCardTray.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import Grid from "@mui/material/Grid2";
-import Resources from "../_subcomponents/PlayerOpponentRows/Resources/Resources";
-import DeckDiscard from "../_subcomponents/PlayerOpponentRows/DeckDiscard/DeckDiscard";
-import CardActionTray from "../_subcomponents/PlayerOpponentRows/CardActionTray/CardActionTray";
+import Resources from "../_subcomponents/PlayerTray/Resources";
+import DeckDiscard from "../_subcomponents/PlayerTray/DeckDiscard";
 import { OpponentCardTrayProps } from "@/app/_components/Gameboard/GameboardTypes";
 
 const OpponentCardTray: React.FC<OpponentCardTrayProps> = ({ participant }) => {
@@ -35,17 +34,10 @@ const OpponentCardTray: React.FC<OpponentCardTrayProps> = ({ participant }) => {
 		<Grid container sx={{ height: "15%" }}>
 			<Grid size={3} sx={leftColumn}>
 				<Resources
-					availableResources={2}
-					totalResources={4}
-					activePlayer={participant.type}
+					trayPlayer={participant.type}
 				/>
 			</Grid>
 			<Grid size={6} sx={centerColumn}>
-				<CardActionTray
-					activePlayer={participant.type}
-					availableCards={participant.cards}
-					// Opponent doesn't need to handle resource selection or playing cards
-				/>
 			</Grid>
 			<Grid size={3} sx={rightColumn}>
 				<DeckDiscard deckSize={participant.deckSize} />
